+++ resources/js/components/KinetixSubscriptionStatus.vue
@@ -41,6 +41,10 @@ const formattedTrialEndsAt = computed(() => {
 
     return new Date(props.subscription.trialEndsAt).toLocaleDateString();
 });
+
+const trialPlanName = computed(() => {
+    return props.subscription?.trialPlan ?? null;
+});
 </script>
 
 <template>
@@ -80,11 +84,21 @@ const formattedTrialEndsAt = computed(() => {
                     v-if="subscription.onTrial && formattedTrialEndsAt"
                     class="p-3 text-xs border-amber-500/20 bg-amber-500/5 text-amber-600 dark:text-amber-400 font-medium rounded-md border"
                 >
-                    {{
-                        t('kinetix.billing_trial_active', {
-                            date: formattedTrialEndsAt,
-                        })
-                    }}
+                    <template v-if="subscription.onGenericTrial && trialPlanName">
+                        {{
+                            t('kinetix.billing_trial_active_plan', {
+                                plan: trialPlanName,
+                                date: formattedTrialEndsAt,
+                            })
+                        }}
+                    </template>
+                    <template v-else>
+                        {{
+                            t('kinetix.billing_trial_active', {
+                                date: formattedTrialEndsAt,
+                            })
+                        }}
+                    </template>
                 </div>
 
                 <div